refactor: remove processImport from the public API
